fix(meet): complete Cloudflare migration validation
