+++ README.md
@@ -21,12 +21,12 @@ CLI, documentation, local services, and deployment tooling behind that system.
 <!-- tanstack-rust-migration-progress:start -->
 _Generated from `apps/tanstack-web/migration/route-manifest.json`. Refresh with `bun migration:tanstack:readme` after route ownership changes._
 
-![Overall migration progress](https://img.shields.io/static/v1?color=fb8c00&label=Overall&message=28%25+terminal&style=flat-square) ![Rust backend migration progress](https://img.shields.io/static/v1?color=cf222e&label=Rust+backend&message=13.28%25+terminal&style=flat-square) ![TanStack Start migration progress](https://img.shields.io/static/v1?color=d29922&label=TanStack+Start&message=67.26%25+terminal&style=flat-square)
+![Overall migration progress](https://img.shields.io/static/v1?color=fb8c00&label=Overall&message=27.96%25+terminal&style=flat-square) ![Rust backend migration progress](https://img.shields.io/static/v1?color=cf222e&label=Rust+backend&message=13.26%25+terminal&style=flat-square) ![TanStack Start migration progress](https://img.shields.io/static/v1?color=d29922&label=TanStack+Start&message=67.26%25+terminal&style=flat-square)
 
 | Track | Progress | Terminal | Migrated | Removed | Remaining |
 | --- | --- | ---: | ---: | ---: | ---: |
-| Overall | `[######--------------]` 28% | 229 / 818 | 215 | 14 | 589 |
-| Rust backend | `[###-----------------]` 13.28% | 79 / 595 | 67 | 12 | 516 |
+| Overall | `[######--------------]` 27.96% | 229 / 819 | 215 | 14 | 590 |
+| Rust backend | `[###-----------------]` 13.26% | 79 / 596 | 67 | 12 | 517 |
 | TanStack Start | `[#############-------]` 67.26% | 150 / 223 | 148 | 2 | 73 |
 
 <details>
@@ -34,7 +34,7 @@ _Generated from `apps/tanstack-web/migration/route-manifest.json`. Refresh with 
 
 | Kind | Progress | Terminal | Remaining |
 | --- | --- | ---: | ---: |
-| api | `[##--------------]` 12.71% | 74 / 582 | 508 |
+| api | `[##--------------]` 12.69% | 74 / 583 | 509 |
 | page | `[###########-----]` 66.45% | 101 / 152 | 51 |
 | layout | `[###########-----]` 70.15% | 47 / 67 | 20 |
 | cron | `[####------------]` 25% | 2 / 8 | 6 |
