+++ apps/docs/platform/architecture/tanstack-rust-migration-plan.mdx
@@ -22,16 +22,16 @@ source of truth. The latest local snapshot on 2026-08-03 is:
 
 | Status | Count |
 | --- | ---: |
-| `legacy-next` | 589 |
+| `legacy-next` | 590 |
 | `migrated` | 215 |
 | `accepted-removal` | 14 |
-| Total tracked artifacts | 818 |
+| Total tracked artifacts | 819 |
 
 Remaining legacy artifacts are concentrated in API/backend ownership:
 
 | Kind | Remaining legacy artifacts |
 | --- | ---: |
-| API | 508 |
+| API | 509 |
 | Cron | 6 |
 | Error boundary | 1 |
 | Loading boundary | 1 |
