+++ apps/docs/platform/architecture/tanstack-rust-migration.mdx
@@ -423,10 +423,10 @@ Current inventory:
 
 - `152` pages
 - `67` layouts
-- `595` route handler artifacts
+- `596` route handler artifacts
 - `8` cron handlers
-- `818` total tracked route artifacts
-- `215` migrated artifacts, `229` terminal artifacts, and `589` remaining
+- `819` total tracked route artifacts
+- `215` migrated artifacts, `229` terminal artifacts, and `590` remaining
   legacy-owned artifacts in the checked manifest
 
 Regenerate it after adding or removing legacy routes:
