+++ apps/tanstack-web/migration/route-manifest.json
@@ -6,11 +6,11 @@
         "acceptedRemoval": 12,
         "key": "api",
         "label": "api",
-        "legacyNext": 508,
-        "migrated": 63,
-        "percentComplete": 12.86,
-        "remaining": 508,
-        "terminal": 75,
+        "legacyNext": 509,
+        "migrated": 62,
+        "percentComplete": 12.69,
+        "remaining": 509,
+        "terminal": 74,
         "total": 583,
         "unknownStatus": 0
       },
@@ -116,11 +116,11 @@
         "acceptedRemoval": 12,
         "key": "rust-backend",
         "label": "Rust backend",
-        "legacyNext": 516,
-        "migrated": 68,
-        "percentComplete": 13.42,
-        "remaining": 516,
-        "terminal": 80,
+        "legacyNext": 517,
+        "migrated": 67,
+        "percentComplete": 13.26,
+        "remaining": 517,
+        "terminal": 79,
         "total": 596,
         "unknownStatus": 0
       },
@@ -323,11 +323,11 @@
       "acceptedRemoval": 14,
       "key": "total",
       "label": "All route artifacts",
-      "legacyNext": 589,
-      "migrated": 216,
-      "percentComplete": 28.08,
-      "remaining": 589,
-      "terminal": 230,
+      "legacyNext": 590,
+      "migrated": 215,
+      "percentComplete": 27.96,
+      "remaining": 590,
+      "terminal": 229,
       "total": 819,
       "unknownStatus": 0
     }
@@ -3696,7 +3696,7 @@
       "methods": ["GET"],
       "routePath": "/api/v1/workspaces/:wsId/meetings",
       "sourceFile": "apps/web/src/app/api/v1/workspaces/[wsId]/meetings/route.ts",
-      "status": "migrated",
+      "status": "legacy-next",
       "targetOwner": "rust-backend",
       "method": "GET",
       "parentId": "api:/api/v1/workspaces/:wsId/meetings:apps/web/src/app/api/v1/workspaces/[wsId]/meetings/route.ts"
